src/test/java/com/daumkakao/localcontents/test/websupport/model/UserTest.java
package com.daumkakao.localcontents.test.websupport.model;

import model.Level;
import model.User;
import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

/**
 * Created by illy on 2016. 1. 8..
 */
public class UserTest {
    User user;

    @Before
    public void setUp(){
        this.user = new User();
    }

    @Test
    public void upgradeLevel(){
        Level[] levels = Level.values();
        for (Level level : levels) {
            if(level.nextLevel() == null) continue;
            user.setLevel(level);
            user.upgradeLevel();
            assertThat(user.getLevel(), is(level.nextLevel()));
        }
    }

    @Test(expected = IllegalStateException.class)
    public void cannotUpgradeLevel(){
        Level[] levels = Level.values();
        for (Level level : levels) {
            if(level.nextLevel() != null) continue;
            user.setLevel(level);
            user.upgradeLevel();
        }
    }
}
